+++ nzshm_model/logic_tree/source_logic_tree/version2/logic_tree.py
@@ -101,6 +101,11 @@ class SourceBranch(Branch):
         """
         return str(self.values)
 
+    @property
+    def registry_identity(self):
+        nrmls = sorted([s.nrml_id for s in self.sources])
+        return "|".join(nrmls)
+
 
 # TODO: protect from users changing tectonic_region_types
 @dataclass